pyproject.toml
[project]
name = "gma2-mcp"
version = "0.1.0"
description = "MCP server for controlling grandMA2 lighting consoles via Telnet"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.21.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]

[project.scripts]
gma2-mcp = "src.server:main"
